Split out code from mirrors_countme/__init__.py
Fixes: #39
